Write skeleton of the specification for the HotStuff subprotocol. Specify `enter_view`, `on_receive_proposal`, and `on_receive_nudge`. Rename `hotstuff::voting` to `roles` and revamp itemdocs in the module.
